Improve the NIO Socket workaround.
Previously all calls to socket.getOutputStream() had to be changed to
SocketUtils.getOutputStream(socket), and anywhere a Socket was wrapped
in an SSLSocket a call to SocketUtils.fixSslOutputStream() had to be
inserted, both of which could be easy to forget.

The new approach wraps Socket and ServerSocket instead, forwarding most
calls to the actual implementations and overriding getOutputStream() to
implement the workaround. This makes everything automatic, as long as
new WorkaroundNio(Server)Socket(ch) is used instead of ch.socket().